+++ R/adjacency.R
@@ -69,16 +69,17 @@ graph.adjacency.dense <- function(adjmatrix, mode=c("directed", "undirected", "m
   res
 }
 
-m2triplet <- get0("mat2triplet", asNamespace("Matrix"), inherits=FALSE)
-if(!is.function(m2triplet)) ## <==> (packageVersion("Matrix") < "1.3")
-    m2triplet <- function(x) {  ## a simplified version of new Matrix' mat2triplet()
-        T <- as(x, "TsparseMatrix")
-        if(is(T, "nsparseMatrix"))
-             list(i = T@i + 1L, j = T@j + 1L)
-        else list(i = T@i + 1L, j = T@j + 1L, x = T@x)
-    }
+## helper function to replace Matrix::summary() in a way that ensures that we
+## have a third column even when Matrix::summary() returned the non-zero
+## cell coordinates only
+mysummary <- function(x) {
+  result <- Matrix::summary(x)
+  if (ncol(result) < 3) {
+    result <- cbind(result, 1)
+  }
+  result
+}
 
-mysummary <- function(x) do.call(cbind, m2triplet(x))
 
 graph.adjacency.sparse <- function(adjmatrix, mode=c("directed", "undirected", "max",
                                                 "min", "upper", "lower", "plus"),
@@ -94,8 +95,6 @@ graph.adjacency.sparse <- function(adjmatrix, mode=c("directed", "undirected", "
       stop("invalid value supplied for `weighted' argument, please see docs.")
     }
   }
-
-  mysummary <- Matrix::summary
 
   if (nrow(adjmatrix) != ncol(adjmatrix)) {
     stop("not a square matrix")